tenant in separate package
