+++ broker/patron_request/api/api-handler_test.go
@@ -13,13 +13,13 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/indexdata/crosslink/broker/api"
 	"github.com/indexdata/crosslink/broker/common"
 	"github.com/indexdata/crosslink/broker/events"
 	"github.com/indexdata/crosslink/broker/handler"
 	pr_db "github.com/indexdata/crosslink/broker/patron_request/db"
 	"github.com/indexdata/crosslink/broker/patron_request/proapi"
 	prservice "github.com/indexdata/crosslink/broker/patron_request/service"
+	"github.com/indexdata/crosslink/broker/tenant"
 	"github.com/indexdata/crosslink/broker/test/mocks"
 	"github.com/indexdata/crosslink/iso18626"
 	"github.com/jackc/pgx/v5"
@@ -62,7 +62,7 @@ func TestGetDbText(t *testing.T) {
 }
 
 func TestGetPatronRequests(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	params := proapi.GetPatronRequestsParams{
@@ -75,7 +75,7 @@ func TestGetPatronRequests(t *testing.T) {
 }
 
 func TestGetPatronRequestsNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	params := proapi.GetPatronRequestsParams{
@@ -87,7 +87,7 @@ func TestGetPatronRequestsNoSymbol(t *testing.T) {
 }
 
 func TestGetPatronRequestsWithLimits(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	offset := proapi.Offset(10)
@@ -107,7 +107,7 @@ func TestGetPatronRequestsWithLimits(t *testing.T) {
 
 func TestGetPatronRequestsWithRequesterReqId(t *testing.T) {
 	repo := new(PrRepoCapture)
-	handler := NewPrApiHandler(repo, mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(repo, mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	requesterReqID := "req-123"
@@ -126,7 +126,7 @@ func TestGetPatronRequestsWithRequesterReqId(t *testing.T) {
 }
 
 func TestPostPatronRequests(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	id := "1"
 	toCreate := proapi.CreatePatronRequest{
 		Id:              &id,
@@ -145,7 +145,7 @@ func TestPostPatronRequests(t *testing.T) {
 }
 
 func TestPostPatronRequestsMissingSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	toCreate := proapi.PatronRequest{Id: "1"}
 	jsonBytes, err := json.Marshal(toCreate)
 	assert.NoError(t, err, "failed to marshal patron request")
@@ -159,7 +159,7 @@ func TestPostPatronRequestsMissingSymbol(t *testing.T) {
 }
 
 func TestPostPatronRequestsInvalidJson(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", bytes.NewBuffer([]byte("a\": v\"")))
 	rr := httptest.NewRecorder()
 	tenant := proapi.Tenant("test-lib")
@@ -169,7 +169,7 @@ func TestPostPatronRequestsInvalidJson(t *testing.T) {
 }
 
 func TestPostPatronRequestsInvalidIllRequestShape(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	jsonBytes := []byte(`{
 		"id":"1",
 		"requesterSymbol":"` + symbol + `",
@@ -185,7 +185,7 @@ func TestPostPatronRequestsInvalidIllRequestShape(t *testing.T) {
 }
 
 func TestDeletePatronRequestsIdNotFound(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.DeletePatronRequestsId(rr, req, "2", proapi.DeletePatronRequestsIdParams{Symbol: &symbol})
@@ -193,7 +193,7 @@ func TestDeletePatronRequestsIdNotFound(t *testing.T) {
 }
 
 func TestDeletePatronRequestsIdMissingSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.DeletePatronRequestsId(rr, req, "2", proapi.DeletePatronRequestsIdParams{})
@@ -202,7 +202,7 @@ func TestDeletePatronRequestsIdMissingSymbol(t *testing.T) {
 }
 
 func TestDeletePatronRequestsIdError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.DeletePatronRequestsId(rr, req, "1", proapi.DeletePatronRequestsIdParams{Symbol: &symbol})
@@ -211,7 +211,7 @@ func TestDeletePatronRequestsIdError(t *testing.T) {
 }
 
 func TestDeletePatronRequestsId(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.DeletePatronRequestsId(rr, req, "3", proapi.DeletePatronRequestsIdParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -220,7 +220,7 @@ func TestDeletePatronRequestsId(t *testing.T) {
 }
 
 func TestDeletePatronRequestsIdDeleted(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.DeletePatronRequestsId(rr, req, "4", proapi.DeletePatronRequestsIdParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -228,7 +228,7 @@ func TestDeletePatronRequestsIdDeleted(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdMissingSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsId(rr, req, "2", proapi.GetPatronRequestsIdParams{})
@@ -237,7 +237,7 @@ func TestGetPatronRequestsIdMissingSymbol(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdNotFound(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsId(rr, req, "2", proapi.GetPatronRequestsIdParams{Symbol: &symbol})
@@ -245,7 +245,7 @@ func TestGetPatronRequestsIdNotFound(t *testing.T) {
 }
 
 func TestGetPatronRequestsId(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsId(rr, req, "1", proapi.GetPatronRequestsIdParams{Symbol: &symbol})
@@ -254,7 +254,7 @@ func TestGetPatronRequestsId(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdActions(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdActions(rr, req, "3", proapi.GetPatronRequestsIdActionsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -263,7 +263,7 @@ func TestGetPatronRequestsIdActions(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdActionsNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdActions(rr, req, "3", proapi.GetPatronRequestsIdActionsParams{Side: &proapiBorrowingSide})
@@ -272,7 +272,7 @@ func TestGetPatronRequestsIdActionsNoSymbol(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdActionsDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdActions(rr, req, "1", proapi.GetPatronRequestsIdActionsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -281,7 +281,7 @@ func TestGetPatronRequestsIdActionsDbError(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdActionsNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdActions(rr, req, "3", proapi.GetPatronRequestsIdActionsParams{Symbol: &symbol, Side: &proapiLendingSide})
@@ -290,7 +290,7 @@ func TestGetPatronRequestsIdActionsNotFoundBecauseOfSide(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdActionNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdAction(rr, req, "3", proapi.PostPatronRequestsIdActionParams{Side: &proapiBorrowingSide})
@@ -299,7 +299,7 @@ func TestPostPatronRequestsIdActionNoSymbol(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdActionDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdAction(rr, req, "1", proapi.PostPatronRequestsIdActionParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -308,7 +308,7 @@ func TestPostPatronRequestsIdActionDbError(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdActionNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdAction(rr, req, "3", proapi.PostPatronRequestsIdActionParams{Symbol: &symbol, Side: &proapiLendingSide})
@@ -317,7 +317,7 @@ func TestPostPatronRequestsIdActionNotFoundBecauseOfSide(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdActionErrorParsing(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", strings.NewReader("{"))
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdAction(rr, req, "3", proapi.PostPatronRequestsIdActionParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -326,7 +326,7 @@ func TestPostPatronRequestsIdActionErrorParsing(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdEventsNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdEvents(rr, req, "3", proapi.GetPatronRequestsIdEventsParams{Side: &proapiBorrowingSide})
@@ -335,7 +335,7 @@ func TestGetPatronRequestsIdEventsNoSymbol(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdEventsDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdEvents(rr, req, "1", proapi.GetPatronRequestsIdEventsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -344,7 +344,7 @@ func TestGetPatronRequestsIdEventsDbError(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdEventsNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdEvents(rr, req, "3", proapi.GetPatronRequestsIdEventsParams{Symbol: &symbol, Side: &proapiLendingSide})
@@ -353,7 +353,7 @@ func TestGetPatronRequestsIdEventsNotFoundBecauseOfSide(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdEventsErrorGettingEvents(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdEvents(rr, req, "3", proapi.GetPatronRequestsIdEventsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -362,7 +362,7 @@ func TestGetPatronRequestsIdEventsErrorGettingEvents(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdNotificationsNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdNotifications(rr, req, "3", proapi.GetPatronRequestsIdNotificationsParams{Side: &proapiBorrowingSide})
@@ -371,7 +371,7 @@ func TestGetPatronRequestsIdNotificationsNoSymbol(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdNotificationsDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdNotifications(rr, req, "1", proapi.GetPatronRequestsIdNotificationsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -380,7 +380,7 @@ func TestGetPatronRequestsIdNotificationsDbError(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdNotificationsNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdNotifications(rr, req, "3", proapi.GetPatronRequestsIdNotificationsParams{Symbol: &symbol, Side: &proapiLendingSide})
@@ -389,7 +389,7 @@ func TestGetPatronRequestsIdNotificationsNotFoundBecauseOfSide(t *testing.T) {
 }
 
 func TestGetPatronRequestsIdNotificationsErrorGettingEvents(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.GetPatronRequestsIdNotifications(rr, req, "3", proapi.GetPatronRequestsIdNotificationsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -423,7 +423,7 @@ func TestGetPatronRequestsIdNotificationsWithKindFilter(t *testing.T) {
 		},
 		fullCount: 1,
 	}
-	handler := NewPrApiHandler(repo, mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(repo, mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("GET", "/", nil)
 	rr := httptest.NewRecorder()
 
@@ -458,7 +458,7 @@ func TestGetPatronRequestsIdNotificationsWithKindFilter(t *testing.T) {
 }
 
 func TestParseAndValidateIllRequestAndBuildDbPatronRequest(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	ctx := common.CreateExtCtxWithArgs(context.Background(), &common.LoggerArgs{})
 	creationTime := time.Now()
 	id := uuid.NewString()
@@ -485,7 +485,7 @@ func TestParseAndValidateIllRequestAndBuildDbPatronRequest(t *testing.T) {
 }
 
 func TestParseAndValidateIllRequestInvalidRequesterSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	ctx := common.CreateExtCtxWithArgs(context.Background(), &common.LoggerArgs{})
 	invalidSymbol := "REQ"
 
@@ -495,7 +495,7 @@ func TestParseAndValidateIllRequestInvalidRequesterSymbol(t *testing.T) {
 }
 
 func TestParseAndValidateIllRequestInvalidBrokerSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	ctx := common.CreateExtCtxWithArgs(context.Background(), &common.LoggerArgs{})
 	previousBrokerSymbol := brokerSymbol
 	brokerSymbol = "BROKER"
@@ -512,7 +512,7 @@ func TestParseAndValidateIllRequestInvalidBrokerSymbol(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdNotificationsNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdNotifications(rr, req, "3", proapi.PostPatronRequestsIdNotificationsParams{Side: &proapiBorrowingSide})
@@ -521,7 +521,7 @@ func TestPostPatronRequestsIdNotificationsNoSymbol(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdNotificationsDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	body := "{\"note\": \"Say hello\"}"
 	req, _ := http.NewRequest("POST", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -531,7 +531,7 @@ func TestPostPatronRequestsIdNotificationsDbError(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdNotificationsNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	body := "{\"note\": \"Say hello\"}"
 	req, _ := http.NewRequest("POST", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -541,7 +541,7 @@ func TestPostPatronRequestsIdNotificationsNotFoundBecauseOfSide(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdNotificationsErrorSavingNotification(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"note\": \"Say hello\"}"
 	req, _ := http.NewRequest("POST", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -551,7 +551,7 @@ func TestPostPatronRequestsIdNotificationsErrorSavingNotification(t *testing.T) 
 }
 
 func TestPostPatronRequestsIdNotificationsErrorBecauseOfBodyMissing(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("POST", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PostPatronRequestsIdNotifications(rr, req, "3", proapi.PostPatronRequestsIdNotificationsParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -560,7 +560,7 @@ func TestPostPatronRequestsIdNotificationsErrorBecauseOfBodyMissing(t *testing.T
 }
 
 func TestPostPatronRequestsIdNotificationsErrorBecauseOfBody(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"note"
 	req, _ := http.NewRequest("POST", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -570,7 +570,7 @@ func TestPostPatronRequestsIdNotificationsErrorBecauseOfBody(t *testing.T) {
 }
 
 func TestPostPatronRequestsIdNotificationsErrorFailedSendOnlyLogged(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositorySuccess), *api.NewTenantContext(), new(MockIso18626Handler), 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositorySuccess), *tenant.NewContext(), new(MockIso18626Handler), 10)
 	body := "{\"note\": \"Say hello\"}"
 	req, _ := http.NewRequest("POST", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -580,7 +580,7 @@ func TestPostPatronRequestsIdNotificationsErrorFailedSendOnlyLogged(t *testing.T
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNoSymbol(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("PUT", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PutPatronRequestsIdNotificationsNotificationIdReceipt(rr, req, "3", "n1", proapi.PutPatronRequestsIdNotificationsNotificationIdReceiptParams{Side: &proapiBorrowingSide})
@@ -589,7 +589,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNoSymbol(t *testin
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptDbError(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -599,7 +599,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptDbError(t *testing
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNotFoundBecauseOfSide(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, mockEventRepo, *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -609,7 +609,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNotFoundBecauseOfS
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorReadingNotification(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -619,7 +619,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorReadingNotifi
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNotFound(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -629,7 +629,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptNotFound(t *testin
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorBecauseOfBodyMissing(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	req, _ := http.NewRequest("PUT", "/", nil)
 	rr := httptest.NewRecorder()
 	handler.PutPatronRequestsIdNotificationsNotificationIdReceipt(rr, req, "3", "n1", proapi.PutPatronRequestsIdNotificationsNotificationIdReceiptParams{Symbol: &symbol, Side: &proapiBorrowingSide})
@@ -638,7 +638,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorBecauseOfBody
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorBecauseOfBody(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"receipt"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -648,7 +648,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptErrorBecauseOfBody
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptPrDoesNotOwn(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
@@ -658,7 +658,7 @@ func TestPutPatronRequestsIdNotificationsNotificationIdReceiptPrDoesNotOwn(t *te
 }
 
 func TestPutPatronRequestsIdNotificationsNotificationIdReceiptFailedToSave(t *testing.T) {
-	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *api.NewTenantContext(), nil, 10)
+	handler := NewPrApiHandler(new(PrRepoError), mockEventBus, new(mocks.MockEventRepositoryError), *tenant.NewContext(), nil, 10)
 	body := "{\"receipt\": \"SEEN\"}"
 	req, _ := http.NewRequest("PUT", "/", bytes.NewBufferString(body))
 	rr := httptest.NewRecorder()
